backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-cloud-speech==2.24.0
google-cloud-translate==3.15.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0
pydub==0.25.1
motor==3.3.2
transformers==4.36.2
torch>=2.10.0
sentencepiece==0.1.99
dnspython
scikit-learn
accelerate
